Added 'Operations' feature
The 'Menu Navigation' feature was integrated with a new one:

Now you can write numbers on screen (decimals too). If you push 'D' button, you can open the MENU and select the sign you want to use in your operations, putting it on the screen.

'#' button is an equal. If you push that button, the result appears on screen.

***Minor changes made. TODOs at the end of the code.