Added 'Hard Operations' feature
It is considered a 'Hard operation':

power, logarithms, square roots, sine, cosine and tangent.

They are all now working in the calculator. You can use them from the menu as you wish.

It was also added a button to delete the last character written ('A' button), to put a ')' ('B' button) and to put a ',' ('C' button).